the_bot_bender/requirements.txt
adal==1.2.7
botbuilder-integration-aiohttp>=4.14.0
aiounittest==1.3.0
async-timeout==3.0.1
attrs==22.2.0
azure-cognitiveservices-language-luis==0.2.0
azure-common==1.1.28
azure-core==1.26.2
azure-cosmos==3.2.0
azure-mgmt-core==1.3.2
azure-mgmt-cosmosdb==6.0.0
azure-storage-blob==12.7.0
azure-storage-queue==12.1.5
Babel==2.9.1
botbuilder-ai>=4.14.0
botbuilder-azure>=4.14.0
botbuilder-core>=4.14.0
botbuilder-dialogs>=4.14.0
botbuilder-integration-aiohttp>=4.14.0
botbuilder-schema==4.14.3
botframework-connector==4.14.3
botframework-streaming==4.14.3
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.0.1
cryptography==39.0.0
datedelta==1.4
Dialogs==1.0
emoji==1.7.0
grapheme==0.6.0
idna==3.4
isodate==0.6.1
jsonpickle==1.4.2
msal==1.20.0
msrest==0.6.10
msrestazure==0.6.4
multidict==6.0.4
multipledispatch==0.6.0
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7.1
recognizers-text==1.0.2a2
recognizers-text-choice==1.0.2a2
recognizers-text-date-time==1.0.2a2
recognizers-text-number==1.0.2a2
recognizers-text-number-with-unit==1.0.2a2
regex==2019.8.19
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.4.0
urllib3==1.26.14
yarl==1.4.2
